\name{data}
\docType{data}
\alias{scores}
\alias{events}
\title{Learning Analytics Sandbox Data Sets}
\description{

Currently, two data sets are provided. They contain the events learners have engaged in (from the logfiles of the LMS) as well as the scores of the learners in each of the multiple assessments conducted throughout the course.

The sets can be loaded by calling \code{data(scores)} and \code{data(events)}.
 
}
\usage{
   data(scores)
   data(events)
}
\author{ Abelardo Pardo \email{abel@it.uc3m.es}, Fridolin Wild \email{f.wild@open.ac.uk} }
\format{dataframes}
\keyword{datasets}
\keyword{array}
