\name{env}
\alias{env}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Environmental input file format for \code{\link{LFMM}}}
\description{
description of the \code{env} format for the environmental variable input file for \code{\link{LFMM}}. 
}
\details{
The \code{env} format has one row for each individual. Each row contains one value per environmental variable 
(separated by spaces or tabulations).
}
\author{
    Eric Frichot
}
\seealso{
\code{\link{LFMM}}
\code{\link{lfmm}}
\code{\link{zscore}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{format}
\keyword{LFMM}
