\name{example}
\docType{data}
\alias{example}
\title{Example tutorial data sets}
\description{
This dataset is composed of a genotypic matrix called example.R with 83 individuals for 500 SNPs.
The last 50 SNPs are correlated with an environmental variable called example.C.
This dataset is a subset of the dataset displayed in the note associated with the package.
}
\usage{example}
\keyword{example}
