\name{offset_example}
\docType{data}
\alias{offset_example}
\title{Example data for genetic offset analysis}
\description{
The data set is composed of a genotypic matrix stored in a lfmm format (geno) containing 200 
individuals genotyped at 510 SNPs, a matrix with 4 correlated environmental variables measured for each individual in the env format, and a matrix with the same 4 variables after environmental change (env.pred). 
}
\value{
\item{geno}{
A genotypic matrix that contains haploid genotypes for 200 individuals at 510 SNPs (lfmm format).
}
\item{env}{
A matrix with 4 correlated environmental variables measured for 200 genotyped individuals.
}
\item{env.pred}{
A matrix with the same 4 variables predicted for the 200 individuals after environmental change.
}
}
\usage{offset_example}
\keyword{offset_example}
