\name{LEA-package}
\alias{LEA-package}
\docType{package}
\title{
LEA: Landscape Ecological Association
}
\description{
R package for population genetics.
}
\details{
\tabular{ll}{
Package: \tab LEA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-16\cr
License: \tab GPL-3\cr
}
}
\author{
Eric Frichot
Maintainer: Eric Frichot <eric.frichot@imag.fr>
}
\keyword{ package }
