\name{tuto}
\docType{data}
\alias{tuto}
\title{Main tutorial data sets}
\description{
  R is the genotype matrix. It is composed of 50 individuals and 10000 SNPs.
  C is the variable matrix. It is composed of two variables for 50 individuals.
  The 50 individuals come from 3 ancestral populations. The first 20 individuals are from population 1. The second 10 individuals come form population 2. the third 20 individuals come from population 3.
  Among the 10000 SNPs, the 
}
\usage{tuto}
\keyword{tutorial}
