\name{tutorial}
\docType{data}
\alias{tutorial}
\title{Main tutorial data sets}
\description{
This dataset is composed of a genotypic matrix called tutorial.R with 165 individuals for 1000 SNPs.
The last 100 SNPs are correlated with an environmental variable called tutorial.C.
This dataset is a subset of the dataset displayed in the note associated with the package.
}
\usage{tutorial}
\keyword{tutorial}
