\name{LEA-package}
\alias{LEA-package}
\docType{package}
\title{
LEA: an R package for Landscape and Ecological Associations studies.
}
\description{
LEA is an R package dedicated to landscape genomics and ecological association 
tests. LEA can run analyses of population structure and genome scans for local 
adaptation. It includes statistical methods for estimating ancestry 
coefficients from large genotypic matrices and evaluating the number of 
ancestral populations (snmf, pca) and identifying genetic polymorphisms that 
exhibit high correlation with some environmental gradient or with the variables
used as proxies for ecological pressures (lfmm). LEA is mainly based on 
optimized C programs that can scale with the dimension of very large data sets.
}
\details{
\tabular{ll}{
Package: \tab LEA\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2017-07-16\cr
License: \tab GPL-3\cr
}
}
\author{
Eric Frichot
Olivier Francois
Maintainer: Olivier Francois <olivier.francois@grenoble-inp.fr>
}
\keyword{ package }
