\name{tutorial}
\docType{data}
\alias{tutorial}
\title{Example tutorial data sets}
\description{
This dataset is composed of a genotypic matrix called tutorial.R with 50 
individuals for 400 SNPs. The last 50 SNPs are correlated with an 
environmental variable called tutorial.C. This dataset is a subset of the 
dataset displayed in the note associated with the package.
}
\value{
\item{tutorial.R}{
A genotypic matrix with 50 individuals for 400 SNPs. The last 50 SNPs are 
correlated with an environmental variable called tutorial.C. 
}
\item{tutorial.C}{
An environmental variable for the 50 invdividuals.
}
}
\usage{tutorial}
\keyword{tutorial}
