\name{tutorial}
\docType{data}
\alias{tutorial}
\title{Example tutorial data sets}
\description{
This dataset is composed of a genotypic matrix stored tutorial.R with 50 
individuals genotyped for 400 SNPs. The last 50 SNPs are correlated with an 
environmental variable (tutorial.C). This dataset is a subset of the 
data shown in the computer note associated with the package (Frichot and Francois 2015).
}
\value{
\item{tutorial.R}{
A genotypic matrix for 50 individuals genotyped at 400 SNPs. The last 50 SNPs are correlated with an environmental variable stored in tutorial.C. 
}
\item{tutorial.C}{
An environmental variable for 50 invdividuals.
}
}
\usage{tutorial}
\keyword{tutorial}
