\name{tutorial}
\docType{data}
\alias{tutorial}
\title{Example tutorial data sets}
\description{
This data set is composed of a genotypic matrix stored in tutorial.R with 50 
individuals genotyped at 400 SNPs. The last 50 SNPs are correlated with an 
environmental variable recorded in tutorial.C. The data are a subset of the 
data shown in the computer note associated with the package (Frichot and Francois 2015).
}
\value{
\item{tutorial.R}{
A genotypic matrix for 50 individuals genotyped at 400 SNPs. The last 50 SNPs are correlated with an environmental variable stored in tutorial.C. 
}
\item{tutorial.C}{
An environmental variable measured for 50 individuals.
}
}
\usage{tutorial}
\keyword{tutorial}
