\name{zscore.format}
\alias{zscore.format}

\title{Output file format for \code{\link{lfmm}}}
\description{
Description of the \code{zscore} output format of \code{\link{lfmm}}. 
}
\details{
The \code{zscore} format has one row for each SNP. Each row contains three 
values: The first value is the zscore, the second value is the -log10(pvalue),
the third value is the p-value (separated by spaces or tabulations).
}
\author{
    Eric Frichot
}
\seealso{
\code{\link{lfmm}}
\code{\link{lfmm.data}}
\code{\link{env}}
}


\keyword{format}
