\name{zscore}
\alias{zscore}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Output file format for \code{\link{LFMM}}}
\description{
Description of the \code{zscore} output format of \code{\link{LFMM}}. 
}
\details{
The \code{zscore} format has one row for each SNP. Each row contains three values: The first value is the zscore, the second value is the -log10(pvalue), the third value is the p-value (separated by spaces or tabulations).
}
\author{
    Eric Frichot
}
\seealso{
\code{\link{LFMM}}
\code{\link{lfmm}}
\code{\link{env}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{format}
\keyword{LFMM}
